chore: rename generate wheel to manage dependencies
